+++ tests/bot/presenceRouting.test.ts
@@ -202,6 +202,7 @@ describe("presence routing", () => {
     ["support", {}],
     ["restart", {}],
     ["dev_raid_stop", {}],
+    ["dev_raid_reset", {}],
     ["dev_reset_korchma_round", {}],
     ["dev_reset_monster_rest", {}],
     ["dev_add_bandage", {}],